file diffing with confirmation kinda works, buggy though

+++ gen/__init__.py
@@ -4,7 +4,8 @@ import os
 import sys
 import configparser
 
-from gen.utils import noop
+from gen.differ import Differ
+from gen.utils import user_confirmation
 from gen.prompt import get_system_prompt, get_edit_file_system_prompt
 from gen.providers import (
     Cerebras,
@@ -53,18 +54,29 @@ def process_file(args, file):
     if args.edit:
         system_prompt = get_edit_file_system_prompt()
 
+    differ = Differ(file)
+    # should probably not read the file twice like this
+    file.seek(0)
+
     response = generate(
         system_prompt,
         args,
-        stream_cb=noop if args.edit else output_token,
+        stream_cb=differ.output_diff if args.edit else output_token,
         additional_content=file.read(),
     )
 
     if args.edit:
-        file.seek(0)
-        file.write(response)
-        file.truncate()
-        sys.stdout.write(file.name)
+        doit = args.force
+        if not args.force:
+            doit = user_confirmation('Are you sure?')
+
+        if doit:
+            file.seek(0)
+            file.write(response)
+            file.truncate()
+            sys.stdout.write(file.name)
+        else:
+            sys.stdout.write('No changes written')
 
     sys.stdout.write('\n')
